+++ CHANGELOG.md
@@ -4,7 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
 ## [0.11.2] - 11/xx/2021
+
+### Added
+- Extending `dpctl.device_context` with nested contexts (#678)
 
 ## Fixed
 - Fixed issue #649 about incorrect behavior of `.T` method on sliced arrays (#653)